shop/views.py
from rest_framework import viewsets, permissions
from .models import Brand, Product
from .serializers import ProductSerializer


class ProductViewSet(viewsets.ModelViewSet):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer
