requirements.txt
#Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-yasg==1.21.7
social-auth-app-django==5.4.0
Django==4.2.3
psycopg2-binary==2.9.6
Django==4.2.3
gunicorn==21.2.0
